Reduce

Description

Reductions of a Rabin acceptance condition dependent and independent of the transition system of the automaton.

Reductions Independent of the Transition System

set_reduce :: Ord q => DRA q a l -> DRA q a l

reduction of a pair by (E,F) -> (E, F \ E)

>>> accept $ set_reduce exdra
[(fromList [3],fromList [1,2])
,(fromList [1,2,3],fromList [])
,(fromList [],fromList [1,2])
]

set_sat :: Eq q => DRA q a l -> DRA q a l

remove unsatisfiable pairs (where the final set is empty)

>>> accept  $ set_sat $ set_reduce exdra
[(fromList [3],fromList [1,2]),(fromList [],fromList [1,2])]

set_irredundant :: Ord q => DRA q a l -> DRA q a l

remove redundant pairs

>>> accept $ set_irredundant $ set_sat $ set_reduce exdra
[(fromList [],fromList [1,2])]

Reductions Dependent of the Transition System

data SingletonRabinpair q

type for a Rabin pair whose final set is singleton

Constructors

SingletonRabinpair

Fields

infinite :: q

finite :: Set q

type SingletonDRA q a l = Automat q a l [SingletonRabinpair q]

type for Rabin automata where each final set is singleton

top_reduce :: Ord q => DRA q a l -> DRA q a l

reduce a pair (Rabin automaton has to be compact)

>>> accept$ top_reduce $ exdra2 {accept = [(S.fromList [3],S.fromList [1]),(S.fromList [3],S.fromList [2])]}
[(fromList [],fromList [1]),(fromList [],fromList [2])]

top_sat :: Ord q => DRA q a l -> DRA q a l

remove unsatisfiable pairs (Rabin automaton has to be compact)

>>> accept $ top_sat $ toDRA $ set_split exdra1
[(fromList [2],fromList [1])
,(fromList [3],fromList [2])
,(fromList [1],fromList [2])]

top_irredundant :: Ord q => Automat q a l [SingletonRabinpair q] -> Automat q a l [SingletonRabinpair q]

remove redundant pairs top_irredundant :: (Ord q, Eq l, Ord a) => SingletonDRA q a l -> SingletonDRA q a l

>>> accept $ top_irredundant $ toSDRA $ top_sat $ toDRA $ set_split exdra1
[SingletonRabinpair {infinite = 1, finite = fromList [2]}
,SingletonRabinpair {infinite = 2, finite = fromList [3]}]

Splitting and Combinig Pairs

set_split :: DRA q a l -> SingletonDRA q a l

split final sets into singletons

set_combine :: Ord q => DRA q a l -> DRA q a l

combine pair with same excluded set by taking the union of the final sets

top_split :: Ord q => DRA q a l -> DRA q a l

split pairs to become compact

top_combine :: Ord q => DRA q a l -> DRA q a l

combine pairs in different SCCs (only for strongly compact Rabin automata)

Checks

is_compact :: Ord q => DRA q a l -> Bool

is the Rabin automaton compact? (final sets contained in one SCC)

>>> is_compact exdra2
False

is_scc :: Ord q => DRA q a l -> Bool

is the Rabin automaton strongly compact? (each pair contained in one SCC)

>>> is_scc  exdra1
True

Conversion Between DRA and Singleton DRA

toSDRA :: Show q => DRA q a l -> SingletonDRA q a l

only works for automata where all final sets are singletons

toDRA :: SingletonDRA q a l -> DRA q a l

Example Rabin Automata

exdra :: DRA Int () ()

Example Rabin automaton (acceptance condition from Example 6.6 in the thesis)

 exdra = makeAutomat 
    [(1,()), (2,()), (3,())] 
    [(1,(),1), (1,(),2), (2,(),1), (2,(),2), (1,(),3), (3,(),2)] 
    [1] 
    [ (S.fromList [3], S.fromList [1,2])
    , (S.fromList [1,2,3], S.fromList [2])
    , (S.fromList [], S.fromList [1,2])
    ]

exdra1 :: DRA Int () ()

Example Rabin automaton (acceptance condition taken from Example 6.13 in the thesis)

 exdra1 = makeAutomat 
    [(1,()), (2,()), (3,())] 
    [(1,(),1), (1,(),2), (2,(),1), (2,(),2), (1,(),3), (3,(),2)] 
    [1] 
    [ (S.fromList [2], S.fromList [1])
    , (S.fromList [3], S.fromList [2])
    , (S.fromList [1], S.fromList [2,3])
    ]

exdra2 :: DRA Int () ()

example Rabin automaton (from Example 6.18 in the thesis)

 exdra2 = makeAutomat
    [(1,()), (2,()), (3,())]
    [(1,(),1), (1,(),2), (2,(),2), (1,(),3), (3,(),2)]
    [1]
    [(S.fromList [3], S.fromList [1,2])]